bug fix
When extracting options, we cannot convert each letter to uppercase because there will be an indefinite article "a" in the Answer, which could cause option "a" to be incorrectly recognized. This leads to instability in the early stages of training, and the model tends to maintain the uppercase letters of the options when answering.